Move good utils funcs into utils/
