+++ utils/index.js
@@ -3,3 +3,4 @@ export * from "./general.js";
 export * from "./type_check.js";
 export * from "./array_utils.js";
 export * from "./file_load.js";
+export * from "./gl_utils.js";
